Move worker
